requirements.txt
ultralytics>=8.0.0
pandas
requests
pillow
opencv-python
tqdm
openpyxl
